Correção de bug
Bug tamanho da foto corrigida
