+++ public/css/styles.css
@@ -10,7 +10,6 @@ li {
     margin-top: 20px;
 }
 
-
 img {
     width: 80px;
     float: left;
@@ -18,7 +17,6 @@ img {
 }
 
 strong {
-
     font-size: 20px;
 }
