+++ public/index.html
@@ -5,31 +5,30 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
-    <link rel="stylesheet" href="/public/css/styles.css">
+    <link rel="stylesheet" href="./public/css/styles.css">
     <title>Repositórios do GitHub</title>
 </head>
 <body>
     <div class="container">
         <div class="card-header">
-            <h3 class="card-title">Repositórios GitHub</h3>
+            <h3 class="card-title">Repositórios Github</h3>
         </div>
-
         <div class="card">
             <div class="card-body">
                 <form autocomplete="off">
                     <div class="form-group">
                         <label for="repositorio">Repositório</label>
-                        <input type="text" class="form-control" id="repositorio" placeholder="Repositório do GitHub" autofocus required>
+                        <input type="text" class="form-control" id="repositorio" placeholder="Repositório do Github" autofocus required>
                         <small id="repositorioHelp" class="form-text text-muted">Informe o repositório para adicionar à lista.</small>
                     </div>
-                    <button type="submit" class="btn btn-primary">adicionar</button>
+                    <button type="submit" class="btn btn-primary">Adicionar</button>
                 </form>
             </div>
             <ul class="list-group">
-
             </ul>
         </div>
     </div>
+    
     <script src="bundle.js"></script>
 </body>
 </html>
