--- src/api2.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const api = axios.create({
-    baseURL: 'https://api.github.com',
-});
-
-export default api;
